Move some Executor logic to its base class
